# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reserva Espacios",
        "description": "API para gestionar reservas de espacios",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservas": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener todas las reservas del usuario autenticado",
                "operationId": "8536ca90a9f5ba3709643c0493395478",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reserva"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener reservas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Crear una nueva reserva",
                "operationId": "3b0d25aeb18200cc0887c6cb9d929934",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "Request"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reserva"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al crear la reserva"
                    }
                }
            }
        },
        "/api/reservas/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener una reserva por ID",
                "operationId": "12b2a5f422f7abede6bff94de0d990f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reserva"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Actualizar una reserva existente",
                "operationId": "91fe86a4a06c23ace3ffc53de226b866",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reserva"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al actualizar la reserva"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Eliminar una reserva",
                "operationId": "992ae788935b42ab455ff366e9887ab7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva eliminada correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error al eliminar la reserva"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reserva": {
                "required": [
                    "espacio_id",
                    "user_id",
                    "fecha",
                    "hora_inicio",
                    "hora_fin"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "espacio_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date"
                    },
                    "hora_inicio": {
                        "type": "string",
                        "format": "time"
                    },
                    "hora_fin": {
                        "type": "string",
                        "format": "time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in the format **Bearer &lt;token&gt;**"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}